refactor(memory): share provider-change model comparison
Extract the provider-change comparison helper into a shared observational memory module so the activation paths use one implementation.

This removes duplicated logic between observational-memory.ts and reflector-runner.ts, which makes the provider-change behavior easier to maintain and reduces the risk of the two paths drifting apart.
