+++ packages/memory/src/processors/observational-memory/reflector-runner.ts
@@ -17,6 +17,7 @@ import {
   createObservationStartMarker,
 } from './markers';
 import type { ModelByInputTokens } from './model-by-input-tokens';
+import { didProviderChange } from './model-context';
 import { registerOp, unregisterOp, isOpActiveInProcess } from './operation-registry';
 import {
   buildReflectorSystemPrompt,
@@ -50,27 +51,6 @@ function formatModelContext(provider?: string, modelId?: string): string | undef
 
 function getCurrentModel(model?: ObservationModelContext): string | undefined {
   return formatModelContext(model?.provider, model?.modelId);
-}
-
-/**
- * Compare a model string derived from past messages against the current actor
- * model. Persisted messages from older code paths may carry a bare `modelId`
- * (no `provider/` prefix) while the current actor always formats as
- * `provider/modelId`. If either side is bare, fall back to comparing just the
- * `modelId` part so a missing provider in history doesn't trigger a spurious
- * provider change.
- */
-function didProviderChange(actorModel?: string, lastModel?: string): boolean {
-  if (actorModel === undefined || lastModel === undefined) return false;
-  if (actorModel === lastModel) return false;
-
-  const actorHasSlash = actorModel.includes('/');
-  const lastHasSlash = lastModel.includes('/');
-  if (actorHasSlash && lastHasSlash) return true;
-
-  const actorModelId = actorHasSlash ? actorModel.slice(actorModel.indexOf('/') + 1) : actorModel;
-  const lastModelId = lastHasSlash ? lastModel.slice(lastModel.indexOf('/') + 1) : lastModel;
-  return actorModelId !== lastModelId;
 }
 
 function getLastModelFromMessageList(messageList?: MessageList): string | undefined {
